Reorganize Header and fix import syntax
