+++ dominion/frontend/src/pages/Country.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Typography } from '@material-ui/core';
-import DocumentHead from '../component/DocumentHead/index';
+import DocumentHead from '../component/DocumentHead';
 
 function Country({ url }) {
   return (
